+++ frontends/main/src/app/(site)/articles/page.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { Metadata } from "next"
 import { standardizeMetadata } from "@/common/metadata"
-import { UserArticleListingPage } from "@/app-pages/UserArticles/UserArticleListingPage"
+import { ArticleListingPage } from "@/app-pages/Articles/ArticleListingPage"
 
 export const metadata: Metadata = standardizeMetadata({
   title: "MIT Learn | Articles",
@@ -9,7 +9,7 @@ export const metadata: Metadata = standardizeMetadata({
 })
 
 const Page: React.FC<PageProps<"/articles">> = () => {
-  return <UserArticleListingPage />
+  return <ArticleListingPage />
 }
 
 export default Page
