+++ frontends/main/src/app/(site)/news/[slugOrId]/draft/page.tsx
@@ -1,24 +1,24 @@
 import React from "react"
-import { ArticleDetailPage } from "@/app-pages/Articles/ArticleDetailPage"
+import { WebsiteContentDetail } from "@/app-pages/WebsiteContent/WebsiteContentDetail"
 import { standardizeMetadata } from "@/common/metadata"
 import RestrictedRoute from "@/components/RestrictedRoute/RestrictedRoute"
 import { Permission } from "api/hooks/user"
 
 export const generateMetadata = async () => {
   return standardizeMetadata({
-    title: "Draft Article",
+    title: "Draft News",
   })
 }
 
 const Page: React.FC<PageProps<"/news/[slugOrId]/draft">> = async (props) => {
   const { slugOrId } = await props.params
 
-  // No prefetching for draft articles - the client-side component
+  // No prefetching for draft News - the client-side component
   // will fetch with user authentication
 
   return (
     <RestrictedRoute requires={Permission.ArticleEditor}>
-      <ArticleDetailPage articleId={slugOrId} learningResourceIds={[]} />
+      <WebsiteContentDetail articleId={slugOrId} learningResourceIds={[]} />
     </RestrictedRoute>
   )
 }
