+++ frontends/main/src/page-components/TiptapEditor/index.ts
@@ -1,5 +1,4 @@
 export { NewsEditor } from "./contentTypes/news/NewsEditor"
-export { ArticleEditor as UserArticleEditor } from "./contentTypes/article/ArticleEditor"
 export { WebsiteContentEditor } from "./core/WebsiteContentEditor"
 export type {
   WebsiteContentEditorProps,
